Finally fixes isometric problems, includes win/loss condition, android port
Isometric visuals are now always correct. (Except for the players, but that's a special cause anyway.)

Now the game knows you WIN if you make a "leap of faith" (walk backwards into the air, fall at least one tile, in the arms of the other player)

And it knows you LOSE if the other player enters your field of view

Lastly, I installed and set up everything for porting to android. Running the game through Android Studio on my USB-connected phone works flawlessly.