Mirrors, interactive mirrors, jittering fix, more levels
Mirrors are back again! They are even interactive (can rotate by pressing the corresponding button)!

They only work in the X,Y (2D) plane though. I still need to add mirrors that can reflect in height (Z axis). Will do that today, probably.

Also found a fix for the jittering: due to my weird choice of coordinate system, when moving vertically, you also move X and Y pos. With slight timing differences, this could cause jittering (because your X and Y position were not integers).

So, now I check if it's a vertical movement, and if so, round the values on the X and Y axis!

Added a few levels with the mirrors, will do a lot more when I have vertical mirrors.