New level (with mirrors), extra interface stuff
Starting to look nice! Biggest challenge now is adding the sound effects (what will ever fit with this soundtrack?) and the intro/outro (I'm curious if I can pull off the voice over)

Once that's done, it's some polishing, creating a logo, checking all the levels, and submitting the game!